Move pure functions from Game to helpers
